Move choose_labels to prompt module
