+++ changes/config.py
@@ -1,17 +1,16 @@
 import io
 import os
 import sys
-
-import click
+from os.path import exists, expanduser, expandvars, join, curdir
 from pathlib import Path
 
+import attr
+import click
 import inflection
 import toml
-import attr
 
-import changes
-from changes.models.repository import GitRepository
 from changes.models import BumpVersion
+from changes import prompt
 from .commands import info, note, debug
 
 AUTH_TOKEN_ENVVAR = 'GITHUB_AUTH_TOKEN'
@@ -125,7 +124,7 @@ def configure_labels(github_labels):
         labels_keyed_by_name[label['name']] = label
 
     # TODO: streamlined support for github defaults: enhancement, bug
-    changelog_worthy_labels = choose_labels([
+    changelog_worthy_labels = prompt.choose_labels([
         properties['name']
         for _, properties in labels_keyed_by_name.items()
     ])
@@ -143,64 +142,6 @@ def configure_labels(github_labels):
         described_labels[label_name] = label_properties
 
     return described_labels
-
-
-def choose_labels(alternatives):
-    """
-    Prompt the user select several labels from the provided alternatives.
-
-    At least one label must be selected.
-
-    :param list alternatives: Sequence of options that are available to select from
-    :return: Several selected labels
-    """
-    if not alternatives:
-        raise ValueError
-
-    if not isinstance(alternatives, list):
-        raise TypeError
-
-    choice_map = OrderedDict(
-      ('{}'.format(i), value) for i, value in enumerate(alternatives, 1)
-    )
-    # prepend a termination option
-    input_terminator = '0'
-    choice_map.update({input_terminator: '<done>'})
-    choice_map.move_to_end('0', last=False)
-
-    choice_indexes = choice_map.keys()
-
-    choice_lines = ['{} - {}'.format(*c) for c in choice_map.items()]
-    prompt = '\n'.join((
-        'Select labels:',
-        '\n'.join(choice_lines),
-        'Choose from {}'.format(', '.join(choice_indexes))
-    ))
-
-    user_choices = set()
-    user_choice = None
-
-    while not user_choice == input_terminator:
-        if user_choices:
-            note('Selected labels: [{}]'.format(', '.join(user_choices)))
-
-        user_choice = click.prompt(
-            prompt,
-            type=click.Choice(choice_indexes),
-            default=input_terminator,
-        )
-        done = user_choice == input_terminator
-        new_selection = user_choice not in user_choices
-        nothing_selected = not user_choices
-
-        if not done and new_selection:
-            user_choices.add(choice_map[user_choice])
-
-        if done and nothing_selected:
-            error('Please select at least one label')
-            user_choice = None
-
-    return user_choices
 
 
 # TODO: borg legacy
