+++ tests/test_init.py
@@ -29,7 +29,7 @@ def answer_prompts(mocker):
     ]
 
     prompt = mocker.patch(
-        'changes.config.choose_labels',
+        'changes.config.prompt.choose_labels',
         autospec=True
     )
     prompt.return_value = ['bug']
